Implement recurring event backend for local calendars
- Events have new properties recurrenceRule, parentEvent, recurrenceStartTime, instances
- Instances of recurring events can easily be created from the master event or updated when the master event is changed
- Implement exceptions and exclusions from a series of recurring events
- Implement saving all of the above to the database
- Implement generation of recurring instances

+++ app/logic/Calendar/Event.ts
@@ -1,5 +1,6 @@
 import type { PersonUID } from "../Abstract/PersonUID";
 import type { Calendar } from "./Calendar";
+import type { RecurrenceRule } from "./RecurrenceRule";
 import { ArrayColl } from "svelte-collections";
 import { assert, randomID } from "../util/util";
 import { Observable, notifyChangedProperty } from "../util/Observable";
@@ -27,6 +28,14 @@ export class Event extends Observable {
   @notifyChangedProperty
   repeat = false;
   @notifyChangedProperty
+  recurrenceRule: RecurrenceRule | undefined;
+  @notifyChangedProperty
+  parentEvent: Event | undefined;
+  /** in case a recurrence instance had its start time modified */
+  recurrenceStartTime: Date | undefined;
+  /** null means deleted instance, undefined means not filled yet */
+  readonly instances: Array<Event | null | undefined> = [];
+  @notifyChangedProperty
   alarm: Date = null;
 
   @notifyChangedProperty
@@ -44,10 +53,14 @@ export class Event extends Observable {
   @notifyChangedProperty
   calendar: Calendar;
 
-  constructor(calendar: Calendar) {
+  constructor(calendar: Calendar, parentEvent?: Event) {
     super();
     this.id = randomID();
     this.calendar = calendar;
+    if (parentEvent) {
+      this.parentEvent = parentEvent;
+      this.copyFromParent();
+    }
   }
 
   /** in seconds */
@@ -83,15 +96,79 @@ export class Event extends Observable {
     this.duration = days * 86400;
   }
 
+  copyFromParent() {
+    let parentEvent = this.parentEvent;
+    this.calUID = parentEvent.calUID;
+    this.title = parentEvent.title;
+    this.descriptionText = parentEvent.descriptionText;
+    this.descriptionHTML = parentEvent.descriptionHTML;
+    this.allDay = parentEvent.allDay;
+    this.parentEvent = parentEvent;
+    this.location = parentEvent.location;
+    this.isOnline = parentEvent.isOnline;
+    this.isPresence = parentEvent.isPresence;
+    this.onlineMeetingURL = parentEvent.onlineMeetingURL;
+    this.participants.replaceAll(parentEvent.participants);
+  }
+
   async save() {
     assert(this.calendar, "To save an event, it needs to be in a calendar first");
     assert(this.calendar.storage, "To save an event, the calendar needs to be saved first");
     await this.calendar.storage.saveEvent(this);
+    for (let occurrence of this.instances) {
+      if (occurrence && !occurrence.dbID) {
+        occurrence.copyFromParent();
+      }
+    }
   }
 
   async deleteIt() {
     assert(this.calendar, "To delete an event, it needs to be in a calendar first");
     assert(this.calendar.storage, "To delete an event, the calendar needs to be saved first");
-    await this.calendar.storage.deleteEvent(this);
+    if (this.dbID) {
+      await this.calendar.storage.deleteEvent(this);
+    }
+    this.calendar.events.remove(this);
+    this.calendar.events.removeAll(this.instances);
+    if (this.parentEvent) {
+      let pos = this.parentEvent.instances.indexOf(this);
+      if (pos >= 0) {
+        this.parentEvent.instances[pos] = null;
+        await this.calendar.storage.saveEvent(this.parentEvent);
+      }
+    }
+  }
+
+  clearExceptions() {
+    this.calendar.events.removeAll(this.instances.filter(Boolean));
+    for (let event of this.instances) {
+      if (event?.dbID) {
+        this.calendar.storage.deleteEvent(event);
+      }
+    }
+    this.instances.length = 0;
+  }
+
+  /**
+   * Removes any previous instance at that position from the calendar
+   * (and also database when an exception subsequently becomes an exclusion).
+   */
+  replaceInstance(index: number, occurrence: Event) {
+    let previous = this.instances[index];
+    this.instances[index] = occurrence;
+    // There won't be a previous instance if this is a new recurring event
+    // and we haven't filled its occurrences yet. Or, this might be an update
+    // to a known existing exception.
+    if (!previous || previous == occurrence) {
+      return;
+    }
+    // Three cases remain:
+    // 1. Deleting a filled instance.
+    // 2. Replacing a filled instance with an exception.
+    // 3. Deleting an exception. In this case, also need to delete from db.
+    this.calendar.events.remove(previous);
+    if (previous.dbID) {
+      this.calendar.storage.deleteEvent(previous);
+    }
   }
 }
